scylla_cql/serialize: Extract row tests to a separate file
